Generalized tree class for re-use in other sites
